Add files via upload
PythonFinalV3 is the main and first program.  It prompts users to input their data, which it then converts it to a csv.  PythonFinalp2 takes the graph that was created in PythonToRProgram.R and displays it.  I really need to think of better names the next time I do something like this...